+++ spec/spec_helper_acceptance.rb
@@ -1,6 +1,4 @@
 require 'beaker-rspec'
-require 'tmpdir'
-require 'yaml'
 require 'simp/beaker_helpers'
 include Simp::BeakerHelpers
 
@@ -15,78 +13,6 @@ unless ENV['BEAKER_provision'] == 'no'
   end
 end
 
-# Install all `*.rpm` files in `rpm_staging_dir/`
-def install_rpms_staged_on(hosts,rpm_staging_dir)
-  on hosts, "yum localinstall -y #{rpm_staging_dir}/*.rpm"
-end
-
-# If local .rpm files have been staged in an 'rpms/' directory at the top
-# level of the repository, upload them to the SUTs' RPM staging directories
-def upload_locally_staged_rpms_to(hosts, rpm_staging_dir)
-  rpms = Dir['*.rpm'] + Dir[File.join('rpms','*.rpm')]
-  rpms.each do |f|
-    scp_to(hosts,f,rpm_staging_dir)
-  end
-end
-
-# Implement any workarounds that are needed to get
-def implement_workarounds(hosts)
-  # workaround for dbus config file mismatch error:
-  #
-  # "dbus[562]: [system] Unable to reload configuration: Configuration file
-  #  needs one or more <listen> elements giving addresses"
-  on hosts, 'systemctl restart dbus'
-end
-
-def install_pre_suite_rpms(hosts)
-  rpm_staging_dir = "/root/rpms.#{$$}"
-
-  on hosts, "mkdir -p #{rpm_staging_dir}"
-  upload_locally_staged_rpms_to(hosts, rpm_staging_dir)
-  install_rpms_staged_on(hosts, rpm_staging_dir)
-end
-
-# starts tpm2sim service on (hosts)
-def start_tpm2sim_on(hosts)
-  on hosts, 'systemctl start simp-tpm2-simulator', pty: true, run_in_parallel: true
-end
-
-def config_abrmd_for_tpm2sim_on(hosts)
-  on hosts, 'mkdir -p /etc/systemd/system/tpm2-abrmd.service.d'
-
-  # Configure the TAB/RM to talk to the TPM2 simulator
-  extra_file=<<-SYSTEMD.gsub(/^\s*/,'')
-  [Service]
-  ExecStart=
-  ExecStart=/sbin/tpm2-abrmd -t socket
-  SYSTEMD
-
-  create_remote_file hosts, '/etc/systemd/system/tpm2-abrmd.service.d/override.conf', extra_file
-  on hosts, 'systemctl daemon-reload'
-
-  on hosts, 'systemctl list-unit-files | grep tpm2-abrmd ' \
-    + '&& systemctl restart tpm2-abrmd ' \
-    + %q[|| echo "tpm2-abrmd.service not restarted because it doesn't exist"]
-end
-
-# start the tpm2sim and override tpm2-abrmd's systemd config use it
-# assumes the tpm2sim has been installed on the hosts
-def configure_tpm2_0_tools(hosts)
-  start_tpm2sim_on(hosts)
-  config_abrmd_for_tpm2sim_on(hosts)
-end
-
-#
-# This is a helper to get the status of the TPM so it can be compared against the
-# the expected results.
-
-def get_tpm2_status(host)
-    stdout = on(host, 'facter -p -y tpm2 --strict').stdout
-    fact = YAML.safe_load(stdout)['tpm2']
-    tpm2_status = fact['tpm2_getcap']['properties-variable']['TPM_PT_PERSISTENT']
-    [tpm2_status['ownerAuthSet'],tpm2_status['endorsementAuthSet'],tpm2_status['lockoutAuthSet']]
-end
-
 RSpec.configure do |c|
   # ensure that environment OS is ready on each host
   fix_errata_on hosts
@@ -97,7 +23,6 @@ RSpec.configure do |c|
   # Configure all nodes in nodeset
   c.before :suite do
     begin
-
       # Install modules and dependencies from spec/fixtures/modules
       copy_fixture_modules_to( hosts )
       begin
@@ -114,10 +39,6 @@ RSpec.configure do |c|
 
       # add PKI keys
       copy_keydist_to(server)
-
-      implement_workarounds(hosts)
-      install_pre_suite_rpms(hosts)
-      configure_tpm2_0_tools(hosts)
     rescue StandardError, ScriptError => e
       if ENV['PRY']
         require 'pry'; binding.pry
